Create user/post relation.

+++ src/RestController/PostController.php
@@ -5,6 +5,7 @@ namespace App\RestController;
 use App\Entity\Post;
 use App\Request\Post\CreatePostRequest;
 use App\Request\Post\UpdatePostRequest;
+use App\Security\Actions;
 use App\Service\PostService;
 use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Controller\FOSRestController;
@@ -20,7 +21,7 @@ class PostController extends FOSRestController
 {
     /**
      * @Rest\Post("/", name="create_post")
-     * @Rest\View(serializerGroups={"post_details", "category_list"})
+     * @Rest\View(serializerGroups={"post_details", "category_list", "user_list"})
      *
      * @param CreatePostRequest $postRequest
      * @param PostService $service
@@ -29,14 +30,16 @@ class PostController extends FOSRestController
      */
     public function createPostAction(CreatePostRequest $postRequest, PostService $service): View
     {
-        $post = $service->create($postRequest);
+        $this->denyAccessUnlessGranted(Actions::CREATE, new Post());
+
+        $post = $service->create($postRequest, $this->getUser());
 
         return View::create($post, Response::HTTP_CREATED);
     }
 
     /**
      * @Rest\Put("/{id}", name="update_post")
-     * @Rest\View(serializerGroups={"post_details", "category_list"})
+     * @Rest\View(serializerGroups={"post_details", "category_list", "user_list"})
      * @ParamConverter("post", class="App:Post")
      *
      * @param UpdatePostRequest $postRequest
@@ -47,6 +50,8 @@ class PostController extends FOSRestController
      */
     public function updatePostAction(UpdatePostRequest $postRequest, Post $post, PostService $service): View
     {
+        $this->denyAccessUnlessGranted(Actions::EDIT, $post);
+
         $service->update($postRequest, $post);
 
         return View::create($post, Response::HTTP_OK);
@@ -54,7 +59,7 @@ class PostController extends FOSRestController
 
     /**
      * @Rest\Get("/{id}", name="get_single_post")
-     * @Rest\View(serializerGroups={"post_details", "category_list"})
+     * @Rest\View(serializerGroups={"post_details", "category_list", "user_list"})
      * @ParamConverter("post", class="App:Post")
      *
      * @param Post $post
@@ -78,6 +83,8 @@ class PostController extends FOSRestController
      */
     public function deletePostAction(Post $post, PostService $service): View
     {
+        $this->denyAccessUnlessGranted(Actions::DELETE, $post);
+
         $service->delete($post);
 
         return View::create($post, Response::HTTP_OK);
@@ -85,7 +92,7 @@ class PostController extends FOSRestController
 
     /**
      * @Rest\Get("/", name="get_all_posts")
-     * @Rest\View(serializerGroups={"post_list", "category_list"})
+     * @Rest\View(serializerGroups={"post_list", "category_list", "user_list"})
      *
      * @param PostService $service
      *
